fastapi_app/pyproject.toml
[tool.poetry]
name = "auth-api"
version = "0.1.0"
description = "Authentication API with FastAPI"
authors = ["Devin AI <158243242+devin-ai-integration[bot]@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.115.6"}
uvicorn = "^0.34.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
sqlalchemy = "^2.0.25"
psycopg2-binary = "^2.9.9"
python-multipart = "^0.0.7"
email-validator = "^2.1.0.post1"
aiosmtplib = "^2.0.2"
pydantic = "^2.5.3"
python-dotenv = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
